Add systemd service file (#1)
This adds a systemd service file to automatically start the AniMe matrix.

This leads to a limitation that the Python script cannot access the display and therefore cannot listen to the keyboard shortcut. That limitation is migitated by providing a way to stop and start the systemd service using the ROG button.